addOrder

+++ mai-market/src/main/java/ru/geekbrains/mai/market/maimarket/dtos/ProductDto.java
@@ -6,7 +6,7 @@ import ru.geekbrains.mai.market.maimarket.models.Product;
 
 import javax.validation.constraints.Min;
 import javax.validation.constraints.Size;
-
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -17,7 +17,7 @@ public class ProductDto {
     private String title;
 
     @Min(value = 1, message = "Min price = 1")
-    private int price;
+    private BigDecimal price;
 
     private String categoryTitle;
